% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backup.R
\name{backup_tbl}
\alias{backup_tbl}
\title{Download a tbl from a db connection and write to a csv}
\usage{
backup_tbl(tbl_name, src, output_dir)
}
\description{
Download a tbl from a db connection and write to a csv
}

