% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{convert_diff_to_update_statement}
\alias{convert_diff_to_update_statement}
\title{Convert a summary of diffs into a SQL UPDATE query}
\usage{
convert_diff_to_update_statement(src, tbl_name, primary_key, tbl_diff)
}
\description{
Convert a summary of diffs into a SQL UPDATE query
}
