% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{create_diff_table}
\alias{create_diff_table}
\title{Summarize the changes between two data-frames}
\usage{
create_diff_table(new_rows, ref_rows, primary_key)
}
\arguments{
\item{new_rows}{a data-frame}

\item{ref_rows}{a reference version of the data-frame}

\item{primary_key}{the name of a column which is used to unique identify rows
in the data}
}
\value{
a data-frame with the primary key column(s), and the columns Field,
  OldVersion and NewVersion showing the differences between the two
  data-frames
}
\description{
Summarize the changes between two data-frames
}

