% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{delete_rows_in_table}
\alias{delete_rows_in_table}
\title{Remove records in a table}
\usage{
delete_rows_in_table(src, tbl_name, rows, guard = TRUE, preview = TRUE)
}
\arguments{
\item{src}{a dplyr-managed database connection or a MySQLConnection}

\item{tbl_name}{name of the table to modify}

\item{rows}{a data-frame of rows to remove}

\item{guard}{whether to prevent the delete statement (the default) if it
would delete all rows in the table}

\item{preview}{whether the delete should be performed or just previewed}
}
\value{
TRUE if the delete succeeded.
}
\description{
Remove records in a table
}

