% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dump.R
\name{dump_database}
\alias{dump_database}
\title{Backup a database as a sequence of SQL statements}
\usage{
dump_database(cnf_file, backup_dir, db_name = "l2t")
}
\arguments{
\item{cnf_file}{a MySQL config file}

\item{backup_dir}{location where to save the backed up table}

\item{db_name}{the name of the database to connect to. Defaults to "l2t".}
}
\value{
Nothing. Information about the dump command and output are printed to
  the console
}
\description{
https://dev.mysql.com/doc/refman/5.7/en/mysqldump.html
}

