% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%===\%}
\alias{\%===\%}
\alias{\%!==\%}
\title{A stronger element-wise equality operators}
\usage{
x \%===\% y

x \%!==\% y
}
\arguments{
\item{x, y}{vectors to be compared. They must have the same length.}
}
\value{
\code{\%===\%} returns \code{TRUE} wherever is \code{x[i]} the same as \code{y[i]}
and \code{\%!==\%} returns \code{FALSE} wherever is \code{x[i]} the same as \code{y[i]}.
}
\description{
These operators work like \code{==} or \code{!=} but they handle comparison with \code{NA}
values. \code{1 \%===\% NA} is \code{FALSE} and \code{NA \%===\% NA} is \code{TRUE}.
}
\examples{
1:3  \%===\% 1:3
#> [1] TRUE TRUE TRUE

1:3 \%===\% 4:6
#> [1] FALSE FALSE FALSE

c(NA, NaN) \%===\% c(NA, NaN)
#> [1] TRUE TRUE

c(NA, 1) \%===\% c(1, NA)
#> [1] FALSE FALSE
}
