% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{find_new_rows_in_table}
\alias{find_new_rows_in_table}
\title{Compare two tables to find new records}
\usage{
find_new_rows_in_table(data, ref_data, required_cols,
  extra_cols = character(0))
}
\arguments{
\item{data}{data-frame to check for changes}

\item{ref_data}{a reference data-frame}

\item{required_cols}{the names of columns that both tables to need to have
and which should only contain non-\code{NA} values.}

\item{extra_cols}{the names of additional columns which should have the same
values in each table}
}
\value{
a data-frame of rows of rows that are in \code{data} but not in \code{ref_data}
}
\description{
Compare two tables to find new records
}
