% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{l2t_connect}
\alias{l2t_connect}
\title{Connect to the L2T database using a MySQL config file}
\usage{
l2t_connect(cnf_file, db_name = "l2t")
}
\arguments{
\item{cnf_file}{a MySQL config file}

\item{db_name}{the name of the database to connect to. Defaults to "l2t".}
}
\value{
a dplyr database connection
}
\description{
Connect to the L2T database using a MySQL config file
}
\details{
http://svitsrv25.epfl.ch/R-doc/library/RMySQL/html/RMySQL-package.html
}

