% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect.R
\name{l2t_writer_connect}
\alias{l2t_writer_connect}
\title{Connect to the L2T database with RMySQL}
\usage{
l2t_writer_connect(cnf_file, db_name = "l2t")
}
\arguments{
\item{cnf_file}{a MySQL config file}

\item{db_name}{the name of the database to connect to. Defaults to "l2t".}
}
\value{
a database connection via RMySQL capable of writing to the database.
}
\description{
Connect to the L2T database with RMySQL
}

