% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L2TDatabase.R
\docType{package}
\name{L2TDatabase}
\alias{L2TDatabase}
\alias{L2TDatabase-package}
\title{Helper functions for working with L2T's MySQL database}
\description{
Helper functions for working with L2T's MySQL database
}

