% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{make_cnf_file}
\alias{make_cnf_file}
\title{Create a MySQL config file}
\usage{
make_cnf_file(dest = "./db.cnf", user = "", password = "", host = "",
  port = 3306, db = "")
}
\arguments{
\item{dest}{file-name where to save the .cnf file}

\item{user}{login credential for db connection. Defaults to "".}

\item{password}{login credential for db connection. Defaults to "".}

\item{host}{hostname of the db connection. Defaults to "".}

\item{port}{port for the db connection. Defaults to 3306.}

\item{db}{name of the database. Defaults to "".}
}
\value{
the lines of the populated .cnf template file are returned. These
lines are also written to the file named in \code{dest} and printed to the
console.
}
\description{
Create a MySQL config file
}
