% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_columns}
\alias{match_columns}
\title{Make columns in one data-frame match those in a second data-frame}
\usage{
match_columns(df1, df2)
}
\description{
Make columns in one data-frame match those in a second data-frame
}

