% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{undo_excel_date}
\alias{undo_excel_date}
\title{Convert an Excel date to a POSIX date}
\usage{
undo_excel_date(dates)
}
\arguments{
\item{dates}{a vector of dates (either numeric or character) originating
from an Excel spreadsheet.}
}
\value{
the dates converted to POSIXct objects (see ?DateTimeClasses)
}
\description{
Excel stores dates as an integer representing the number of days since
1/1/1900. Undo that. See
http://www.exceltactics.com/definitive-guide-using-dates-times-excel/
}
\examples{
undo_excel_date("41659")
#> "2014-01-20 UTC"
undo_excel_date(41534)
#> "2013-09-17 UTC"
}

