% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{create_diff_table}
\alias{create_diff_table}
\title{Summarize the changes between two data-frames}
\usage{
create_diff_table(data, ref_data, primary_key)
}
\arguments{
\item{data}{a data-frame}

\item{ref_data}{a reference version of the data-frame}

\item{primary_key}{the name of a column which is used to unique identify rows
in the data}
}
\value{
a data-frame with the primary key column(s), and the columns \code{Field},
\code{OldVersion} and \code{NewVersion} showing the differences between the two
data-frames
}
\description{
Summarize the changes between two data-frames
}
