% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{delete_rows_in_table}
\alias{delete_rows_in_table}
\title{Remove records in a table}
\usage{
delete_rows_in_table(src, tbl_name, rows, guard = TRUE, preview = TRUE)
}
\arguments{
\item{src}{a dplyr-managed database connection or a \code{MySQLConnection}}

\item{tbl_name}{the name of the table to modify}

\item{rows}{a data-frame of rows to remove}

\item{guard}{whether to prevent the delete statement (the default) if it
would delete all rows in the table}

\item{preview}{whether the table-modifying query should be performed or just
previewed}
}
\value{
\code{TRUE} if the query succeeded or if the preview worked without
errors.
}
\description{
Remove records in a table
}
