% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump.R
\name{dump_database}
\alias{dump_database}
\title{Backup a database as a sequence of SQL statements}
\usage{
dump_database(cnf_file, backup_dir, db_name = "l2t")
}
\arguments{
\item{cnf_file}{a MySQL config file}

\item{backup_dir}{location where to save the backed up table}

\item{db_name}{the name of the database to connect to. Defaults to "l2t".}
}
\value{
Nothing. Information about the dump output is printed to the console
}
\description{
This function calls \code{mysqldump} on a database connection described in a
cnf file. Consider reading the
\href{https://dev.mysql.com/doc/refman/5.7/en/mysqldump.html}{documentation
of mysqldump}.
}
\details{
The command-line call uses the \code{--single-transaction} flag because it
doesn't need LOCKED TABLES privilege.
}
