% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/connect.R
\name{l2t_connect}
\alias{l2t_connect}
\title{Connect to the L2T database using a MySQL config file}
\usage{
l2t_connect(cnf_file)
}
\arguments{
\item{cnf_file}{a MySQL config file}
}
\value{
a dplyr database connection
}
\description{
Connect to the L2T database using a MySQL config file
}
\details{
http://svitsrv25.epfl.ch/R-doc/library/RMySQL/html/RMySQL-package.html
}

