% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{overwrite_rows_in_table}
\alias{overwrite_rows_in_table}
\title{Update records in a table}
\usage{
overwrite_rows_in_table(src, tbl_name, rows, preview = TRUE)
}
\arguments{
\item{src}{a dplyr-managed database connection or a MySQLConnection}

\item{tbl_name}{name of the table to update}

\item{rows}{a data-frame of rows with new data}

\item{preview}{whether the update should be performed or just previewed}
}
\value{
TRUE if the update succeeded.
}
\description{
Update records in a table
}

