% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select-helpers}
\alias{select-helpers}
\alias{maybe_starts_with}
\alias{maybe_matches}
\title{Versions of dplyr's select helpers that fail safely}
\usage{
maybe_starts_with(...)

maybe_matches(...)
}
\arguments{
\item{...}{arguments passed on to the dplyr select helpers}
}
\value{
the column indices for matching columns if any are found; otherwise,
an empty numeric vector.
}
\description{
When dplyr's column-selection functions \code{starts_with} or \code{matches} fail to
find a column, no columns are returned. These alternative versions return all
available columns when no match is found. The result is that \code{select(df, V1, maybe_starts_with("V2")} still returns V1 even if there is no V2.
}
